django_projekti/requirements.txt
GitPython
requests
PyExecJS
pymongo
